requirements.txt
asn1crypto==0.24.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
CacheControl==0.12.5
certifi==2019.3.9
cffi==1.11.5
chardet==3.0.4
cloudpickle==0.8.0
cryptography==2.4.2
cycler==0.10.0
git+git://github.com/Cytomine-ULiege/Cytomine-python-client.git
dask==1.1.4
decorator==4.4.0
defusedxml==0.5.0
entrypoints==0.3
future==0.17.1
idna==2.8
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterlab==0.35.4
jupyterlab-server==0.2.0
kiwisolver==1.0.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mistune==0.8.4
msgpack==0.6.1
nbconvert==5.4.1
nbformat==4.4.0
networkx==2.2
notebook==5.7.6
numpy==1.16.2
olefile==0.46
opencv-python==4.0.0.21
opencv-python-headless==4.1.2.30
openslide-python==1.1.1
pandocfilters==1.4.2
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
prometheus-client==0.6.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycosat==0.6.3
pycparser==2.19
pyflakes>=2.1.1
Pygments==2.3.1
pylint>=2.4.4
pyOpenSSL==18.0.0
pyparsing==2.3.1
pyrsistent==0.14.11
PySocks==1.6.8
python-dateutil==2.8.0
pytz==2018.9
PyWavelets==1.0.2
pyzmq==18.0.1
qtconsole==4.4.3
requests>=2.22.0
requests-toolbelt==0.9.1
ruamel.yaml==0.15.46
scikit-image==0.14.2
scipy==1.2.1
Send2Trash==1.5.0
Shapely==1.7b1
six==1.12.0
terminado==0.8.1
testpath==0.4.2
toolz==0.9.0
torch==0.4.1
torchvision==0.2.1
tornado==6.0.1
traitlets==4.3.2
urllib3>=1.25.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
